Rename classes directory to models
