+++ flaskr/models/persona.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, String, Integer, ForeignKey
 from sqlalchemy.orm import relationship
 
-from flaskr.classes.base import Base
+from flaskr.models.base import Base
 
 
 class Persona(Base):
